refactor(interface): 将 Task Models 迁移为 Models 页的 Model Link 内嵌模块;
- 新建 TaskModelLinkSection 组件，从 task-models 页提取任务模型映射逻辑
- 在 Models 页 Vendor Credentials 卡片下方插入 Model Link 区块
- 从 InterfaceNav 移除 Task Models 独立 tab
- 将 /interface/task-models 路由改为重定向到 /interface/models
